+++ package.json
@@ -14,6 +14,7 @@
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-scripts": "4.0.3",
+    "react-select": "^5.4.0",
     "typescript": "^4.2.4",
     "typescript-eslint-parser": "^22.0.0",
     "web-vitals": "^1.1.1"
