+++ src/App.scss
@@ -1,7 +1,12 @@
+* {
+  box-sizing: border-box;
+}
+
 .starter {
   font-size: 18px;
 }
 
 body {
-  margin: 2% 2%;
+  font-family: 'Montserrat', sans-serif;
+  font-weight: 400;
 }
