+++ src/App.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import './App.scss';
-import { Vacancies } from './blocks/Vacancies';
+import { Vacancies } from './blocks/Vacancies/Vacancies';
 
 export const App: React.FC = () => {
   return (
