+++ src/blocks/Vacancies/Vacancies.scss
@@ -1,0 +1,61 @@
+@import '../../utils/mixins.scss';
+
+.field {
+  position: absolute;
+  top: 20px;
+  right: 50px;
+}
+
+.container {
+  width: 76%;
+  margin: 0 auto;
+}
+
+.Vacancies {
+  &__title {
+    font-weight: 900;
+    line-height: 115%;
+    text-decoration-line: underline;
+    font-size: 24px;
+    margin: 35px 0;
+
+    @include on1440 {
+      font-size: 36px;
+      margin: 47px 0;
+    }
+
+    @include on1920 {
+      font-size: 48px;
+      margin: 70px 0;
+    }
+  }
+
+  &__navigation {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    margin-bottom: 85px;
+
+    @include on1440 {
+      margin-bottom: 113px;
+    }
+
+    @include on1920 {
+      margin-bottom: 170px;
+    }
+  }
+
+  &__cards {
+    display: flex;
+    flex-direction: column;
+    gap: 32px;
+
+    @include on1440 {
+      gap: 43px;
+    }
+
+    @include on1920 {
+      gap: 64px;
+    }
+  }
+}
