+++ src/global-styles/search.scss
@@ -1,3 +1,5 @@
+@import '../utils/mixins.scss';
+
 .search {
   &__dropdown {
     display: flex;
@@ -7,5 +9,86 @@
 
   &__dropdown-row {
     width: 100%;
+    border: none;
+    text-align: left;
+    margin-bottom: 1px;
+    background-color: transparent;
+
+    &:hover {
+      cursor: pointer;
+    }
+  }
+
+  &-container {
+    width: 232px;
+    height: 21px;
+
+    @include on1440 {
+      width: 309px;
+      height: 28px;
+    }
+
+    @include on1920 {
+      width: 464px;
+      height: 42px;
+    }
+  }
+
+  &-inner {
+    position: relative;
+    height: 100%;
+  }
+
+  &-input {
+    width: 100%;
+    height: 100%;
+    padding: 0 4%;
+    background: #fff;
+    border: 0.5px solid #000;
+    border-radius: 8px;
+
+    &:focus {
+      outline: none;
+      -webkit-box-shadow: 10px 10px 12px -6px rgba(0,0,0,0.38);
+      -moz-box-shadow: 10px 10px 12px -6px rgba(0,0,0,0.38);
+      box-shadow: 10px 10px 12px -6px rgba(0,0,0,0.38);
+    }
+
+    &::placeholder {
+      text-align: center;
+      font-size: 9px;
+      line-height: 115%;
+      color: #6f6f6f;
+
+      @include on1440 {
+        font-size: 14px;
+      }
+  
+      @include on1920 {
+        font-size: 18px;
+      }
+    }
+  }
+
+  &-icon {
+    position: absolute;
+    top: 0;
+    bottom: 0;
+    margin: auto 0 auto 24%;
+
+    width: 9.5px;
+    height: 9.5px;
+
+    @include on1440 {
+      font-size: 14px;
+      width: 13px;
+      height: 13px;
+    }
+
+    @include on1920 {
+      font-size: 18px;
+      width: 19px;
+      height: 19px;
+    }
   }
 }
